requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mp3play==0.1.15
Pillow==8.1.0
playsound==1.2.2
pygame==2.0.1
pylint==2.6.2
pypiwin32==223
python-dateutil==2.8.1
pywin32==300
requests==2.25.1
six==1.15.0
tkinter-nav==0.0.5
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
win10toast==0.9
wrapt==1.12.1
